Move functions from dp to util and add tests
